Fix Kraken rate-limit handling, currency precision, and Monzo import bugs
Kraken:
- Kraken's cost/fee fields can legitimately exceed a currency's nominal
  decimal places (e.g. a GBP trade leg quoted to 5 decimals). Rather than
  round that away, every currency now shares crypto's storage scale
  (10^18), removing the ISO-4217-based per-currency scale table.
- Add generic, config-driven rate-limit handling (rateLimitMillis,
  rateLimitErrorSubstrings, rateLimitBackoffMillis, maxRateLimitRetries,
  per-endpoint requestCostWeight) so a transient rate-limit error retries
  with backoff instead of abandoning the rest of an endpoint's windows.
- Add Kraken Earn/staking reward ledger endpoints, and
  assetSuffixesToStrip (generic) to resolve Earn-holding asset codes like
  "XETH.F" that were previously silently dropped.
- Fix keyed-object response items (Kraken's TradesHistory/Ledgers) storing
  a fabricated array-index audit path instead of their real object key,
  which broke the "view origin" link.
- Simplify the local strategy-catalog-directory picker from a toggle to a
  direct folder-picker/clear flow.

Monzo:
- MINOR_UNITS_INTEGER amounts (a bank API's raw integer in its own ISO
  4217 minor units, e.g. Monzo's pence) were constructed directly as
  Money at the new, much larger storage scale, silently zeroing every
  amount. Convert through the ISO divisor first (IsoMinorUnitDivisors),
  now overridable per strategy (minorUnitDivisorOverrides) for a provider
  that deviates from the standard.
- Add ApiAccountMappings.staticAccountName so a provider whose account
  "description" isn't a real display name (Monzo's is the account
  holder's own user id) can be named properly, with " Joint" appended
  for a multi-owner account.

+++ app/db/core/src/commonMain/kotlin/com/moneymanager/database/DatabaseConfig.kt
@@ -69,7 +69,7 @@ object DatabaseConfig {
                         assetWriteQueries.insert()
                         assetWriteQueries.lastInsertedId().executeAsOne()
                     }
-                currencyWriteQueries.insert(currencyId, currency.code, currency.displayName, scaleFactor.toLong())
+                currencyWriteQueries.insert(currencyId, currency.code, currency.displayName, scaleFactor)
                 recordSource(
                     DeviceId(WellKnownIds.SYSTEM_DEVICE_ID),
                     EntityType.CURRENCY,
